NP-45337 Allow setting assignee only for open periods
